Added CombatParryingEra and Windows clang support.
- Added: CombatParryingEra ini flag. It controls the parrying behaviour:
   01 = pre-SE parrying chance formula (not using the Bushido skill). Mutually exclusive with 02 flag.
   02 = SE parrying chance formula (using the Bushido skill). Mutually exclusive with 01 flag.
   010 = can parry with a shield
   020 = can parry with a one-handed weapon without a shield
   040 = can parry with two handed weapons
- Fixed: parrying chance was too high if the Parrying skill was > 100.0.

Added another version of CSTypedArray (unfinished yet) using pointers for its elements, instead of storing class instances (which can be a problem if the class is polymorphic). The new approach however destroys data locality, which is a problem to be addressed sooner or later.
Disabled Sphere's call stack for Release builds (and for builds with Address Sanitizer enabled (which should be used exclusively for testing) since it slows down the code).
Fix for a compilation error occurring when MT_NETWORK wasn't active.
Minor fixes suggested by Address Sanitizer.
Renamed sassertion.h to assertion.h. The 's' prefix should be added only to sphere_library code.
Changed push_back with emplace_back in tooltip code to improve performance and allow more compiler optimizations.

+++ src/network/network_multithreaded.cpp
@@ -358,7 +358,7 @@ void NetworkManager::start(void)
 			{
 				// If we have more than one thread (this hasn't sense... at this point isThreaded should be == true)
 				char name[IThread::m_nameMaxLength];
-				sprintf(name, "T_Worker #%u", pThread->getId());
+				sprintf(name, "T_Worker #%u", (uint)pThread->getId());
 				pThread->overwriteInternalThreadName(name);
 			}
